project 1 - ORM playground/backend/requirements/base.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]==3.2.3
django-environ==0.11.2
gunicorn==23.0.0
redis==5.2.0
